+++ src/views/DesignEditor/SingleLayerExport/SingleLayerExport.tsx
@@ -16,15 +16,15 @@ import Paste from "~/components/Icons/Paste"
 import { useEffect, useState } from "react"
 import Elements from "~/components/Icons/Elements"
 import { deviceUploadType } from "~/constants/contants"
-import { DuplicateFunc } from "../utils/tools/DuplicateFunc"
-import { PasteFunc } from "../utils/tools/PasteFunc"
-import { FrontFunc } from "../utils/tools/FrontFunc"
-import { BackFunc } from "../utils/tools/BackFunc"
-import { LockFunc, UnlockFunc } from "../utils/tools/LockUnlockFunc"
-import { UngroupFunc } from "../utils/tools/GroupUngroupFunc"
-import { InvisibleFunc, VisibleFunc } from "../utils/tools/VisibilityFunc"
-import { SetBgFunc } from "../utils/SetBgFunc"
-import { DeleteFunc } from "../utils/tools/DeleteFunc"
+import { DuplicateFunc } from "../utils/functions/tools/DuplicateFunc"
+import { PasteFunc } from "../utils/functions/tools/PasteFunc"
+import { FrontFunc } from "../utils/functions/tools/FrontFunc"
+import { BackFunc } from "../utils/functions/tools/BackFunc"
+import { LockFunc, UnlockFunc } from "../utils/functions/tools/LockUnlockFunc"
+import { UngroupFunc } from "../utils/functions/tools/GroupUngroupFunc"
+import { InvisibleFunc, VisibleFunc } from "../utils/functions/tools/VisibilityFunc"
+import { SetBgFunc } from "../utils/functions/SetBgFunc"
+import { DeleteFunc } from "../utils/functions/tools/DeleteFunc"
 
 const SingleLayerExport = ({ isOpenSlider, activeOb, show }: any) => {
   const contextMenuRequest = useContextMenuRequest()
