+++ src/views/DesignEditor/components/ContextMenu/ContextMenu.tsx
@@ -17,15 +17,15 @@ import DownloadIcon from "~/components/Icons/DownloadIcon"
 import DownloadPopup from "../Footer/Graphic/DownloadPopup/DownloadPopup"
 import Elements from "~/components/Icons/Elements"
 import ContextMenuItem from "./ContextMenuItem"
-import { DuplicateFunc } from "../../utils/tools/DuplicateFunc"
-import { PasteFunc } from "../../utils/tools/PasteFunc"
-import { FrontFunc } from "../../utils/tools/FrontFunc"
-import { BackFunc } from "../../utils/tools/BackFunc"
-import { LockFunc, UnlockFunc } from "../../utils/tools/LockUnlockFunc"
-import { InvisibleFunc, VisibleFunc } from "../../utils/tools/VisibilityFunc"
-import { UngroupFunc } from "../../utils/tools/GroupUngroupFunc"
-import { SetBgFunc } from "../../utils/SetBgFunc"
-import { DeleteFunc } from "../../utils/tools/DeleteFunc"
+import { DuplicateFunc } from "../../utils/functions/tools/DuplicateFunc"
+import { PasteFunc } from "../../utils/functions/tools/PasteFunc"
+import { FrontFunc } from "../../utils/functions/tools/FrontFunc"
+import { BackFunc } from "../../utils/functions/tools/BackFunc"
+import { LockFunc, UnlockFunc } from "../../utils/functions/tools/LockUnlockFunc"
+import { InvisibleFunc, VisibleFunc } from "../../utils/functions/tools/VisibilityFunc"
+import { UngroupFunc } from "../../utils/functions/tools/GroupUngroupFunc"
+import { SetBgFunc } from "../../utils/functions/SetBgFunc"
+import { DeleteFunc } from "../../utils/functions/tools/DeleteFunc"
 
 const ContextMenu = () => {
   const contextMenuRequest = useContextMenuRequest()
