+++ src/views/DesignEditor/utils/functions/SetBgFunc.tsx
