+++ src/views/DesignEditor/utils/functions/tools/DeleteFunc.tsx
