+++ src/views/DesignEditor/utils/functions/tools/DuplicateFunc.tsx
